requirements.txt
streamlit
pyresparser
pdfminer.six
spacy
pandas
nltk
sqlalchemy
